+fonts

+++ app/views/admin_views.py
@@ -1,18 +1,15 @@
+from __future__ import annotations
 from datetime import datetime as dt
-from typing import Union
 from flask import Blueprint, redirect, render_template, request, url_for, current_app as app
 from flask_login import login_required
 from flask_security import login_required, current_user, roles_required
 from app import babel
 import gettext
 
-from ..models.content import ContentType, Genre
-from ..models.tools import Language
-
-
+from ..models.content import Content, ContentType, Genre, Movie, Series
+from ..models.tools import Language, Dictionary
 admin_blueprint = Blueprint('admin', __name__, template_folder='templates')
-dicts: list[Language | ContentType | Genre] = [Language, ContentType, Genre]
-
+dicts = {d.__name__: d for d in Dictionary.__subclasses__()}
 
 @babel.localeselector
 def get_locale():
@@ -32,22 +29,35 @@ def set_lang(lang):
 @admin_blueprint.before_app_first_request
 def init_my_blueprint():
   if not app.user_datastore.get_user('admin@self.com'):
-    en: Language = Language(name='english', code='en').save()
-    ru: Language = Language(name='russian', code='ru').save()
-    en.set_label(en, 'English')
-    en.set_label(ru, 'Английский')
-    ru.set_label(en, 'Russian')
-    ru.set_label(ru, 'Русский')
-    ContentType(name='movies', code='movies')\
-      .set_label(en, 'Movies')\
-      .set_label(ru, 'Фильмы')
-    print(en.label(ru))
     app.user_datastore.create_role(name="admin")
     app.user_datastore.create_user(
       username='admin', email='admin@self.com',
       confirmed_at=dt.now(),
       password='vxtr5w7c_nxd', roles=['admin']
     )
+  if not Language.default:
+    en: Language = Language(code='en_US')
+    ru: Language = Language(code='ru_RU')
+    en.set_title(en, 'English', False)
+    en.set_title(ru, 'Английский', False)
+    ru.set_title(en, 'Russian', False)
+    ru.set_title(ru, 'Русский', False)
+    en.save()
+    ru.save()
+  if not ContentType.objects(code='movies').count():
+    ContentType(code='movies')\
+      .set_title(en, 'Movies', False)\
+      .set_title(ru, 'Фильмы')
+  # movie: Movie = Movie(code='movie-1', type=ContentType.objects.get(code='movies'),
+  # original_language=Language.objects.get(code='en'), original_title='Movie 1')
+  # movie.save()
+  # series: Series = Series(code='series-1', type=ContentType.objects.get(code='movies'),
+  # original_language=Language.objects.get(code='en'), original_title='Series 1')
+  # series.save()
+  
+  # print(Movie.objects.all().to_json())
+  # print(Series.objects.all().to_json())
+  # print(Content.objects.all().to_json())
   pass
 
 
@@ -70,14 +80,7 @@ def admin():
 @login_required
 @roles_required('admin')
 def dictionaries():
-  return render_template('admin/dictionaries.html', active_page='dictionaries', dictionaries=[
-    {'name': _dict.__name__, 'objects': _dict.objects.all().order_by('-active'),
-    'object': _dict, 'fields': _dict._fields}
-    for _dict in dicts
-  ], languages={
-    'en': Language.objects.get(code='en'),
-    'other': Language.objects.all().filter(code__ne='en').order_by('-active')
-  })
+  return render_template('admin/dictionaries.html', dicts=dicts)
 
 
 @admin_blueprint.route('/admin/dictionary/<string:d_name>/add', methods=['POST'])
@@ -85,19 +88,15 @@ def dictionaries():
 @roles_required('admin')
 def dictionary_add(d_name):
   print(request.values)
-  _dict = next(_dict for _dict in dicts if _dict.__name__ == d_name)
-  _object: Language | ContentType | Genre = _dict()
-  after_save_fields = []
+  _dict = dicts[d_name]
+  _object = _dict()
   for key, value in request.values.items():
     if value == '' or key == 'csrf_token': continue
     if key == 'active': value = value in ['on', 'true', 'True', '1', 'yes', 'Yes', True]
-    if key.startswith('label_') or key.startswith('description_'):
-      field, lang = key.split('_')
-      if _dict == Language and lang == 'self':
-        after_save_fields.append((field, value))
-        continue
-      lang = Language.objects.get(code=lang)
-      if field == 'label': _object.set_label(lang, value, False)
+    if key.startswith(('title_', 'description_')):
+      field, lang = key.split('_', 1)
+      lang = _object if lang == 'self' else Language.objects.get(code=lang)
+      if field == 'title': _object.set_title(lang, value, False)
       elif field == 'description': _object.set_description(lang, value, False)
     elif key in _dict._fields:
       setattr(_object, key, value)
@@ -105,27 +104,24 @@ def dictionary_add(d_name):
   try:
     e = None
     _object.save()
-    for field, value in after_save_fields:
-      if field == 'label': _object.set_label(_object, value)
-      elif field == 'description': _object.set_description(_object, value)
   except Exception as e: print(e)
   return redirect(url_for('admin.dictionaries', error=e))
 
 
-@admin_blueprint.route('/admin/dictionary/<string:d_name>/edit/<string:o_id>', methods=['POST'])
+@admin_blueprint.route('/admin/dictionary/<string:d_name>/edit/<string:o_code>', methods=['POST'])
 @login_required
 @roles_required('admin')
-def dictionary_edit(d_name, o_id):
+def dictionary_edit(d_name, o_code):
   print(request.values)
-  _dict = next(_dict for _dict in dicts if _dict.__name__ == d_name)
-  _object: Language | ContentType | Genre = _dict.objects.get(id=o_id)
+  _dict = dicts[d_name]
+  _object = _dict.objects.get(code=o_code)
   for key, value in request.values.items():
     if value == '' or key == 'csrf_token': continue
     if key == 'active': value = value in ['on', 'true', 'True', '1', 'yes', 'Yes', True]
-    if key.startswith('label_') or key.startswith('description_'):
-      field, lang = key.split('_')
-      lang = Language.objects.get(code=lang)
-      if field == 'label': _object.set_label(lang, value, False)
+    if key.startswith(('title_', 'description_')):
+      field, lang = key.split('_', 1)
+      lang = _object if lang == 'self' else Language.objects.get(code=lang)
+      if field == 'title': _object.set_title(lang, value, False)
       elif field == 'description': _object.set_description(lang, value, False)
     elif key in _dict._fields:
       setattr(_object, key, value)
@@ -141,8 +137,9 @@ def dictionary_edit(d_name, o_id):
 @login_required
 @roles_required('admin')
 def dictionary_delete(d_name, o_code):
-  _dict = next(_dict for _dict in dicts if _dict.__name__ == d_name)
-  _object: Language | ContentType | Genre = _dict.objects.get(code=o_code)
+  _dict = dicts[d_name]
+  _object = _dict.objects.get(code=o_code)
+  if _object == Language.default: return redirect(url_for('admin.dictionaries', error='Can\'t delete default language'))
   _object.delete()
   return redirect(url_for('admin.dictionaries'))
